moved del+all to test_dao.py
